# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.Zone:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        example: SE1
        type: string
      timezone:
        example: Europe/Stockholm
        type: string
      updated_at:
        type: string
    required:
    - name
    - timezone
    type: object
host: localhost:8080
info:
  contact: {}
  description: WattWatch API server.
  title: WattWatch API
  version: "1.0"
paths:
...
  /zones:
    get:
      consumes:
      - application/json
      description: Returns a list of all zones
      parameters:
      - description: Search zones by name
        in: query
        name: search
        type: string
      - description: Order by field (name, timezone)
        in: query
        name: order_by
        type: string
      - description: Order descending
        in: query
        name: order_desc
        type: boolean
      - description: Limit results
        in: query
        name: limit
        type: integer
      - description: Offset results
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Zone'
            type: array
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all zones
      tags:
      - zones
    post:
      consumes:
      - application/json
      description: Creates a new zone
      parameters:
      - description: Zone to create
        in: body
        name: zone
        required: true
        schema:
          $ref: '#/definitions/models.Zone'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Zone'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new zone
      tags:
      - zones
  /zones/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an existing zone
      parameters:
      - description: Zone ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid zone ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Zone not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a zone
      tags:
      - zones
    get:
      consumes:
      - application/json
      description: Returns a zone by its ID
      parameters:
      - description: Zone ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Zone'
        "400":
          description: Invalid zone ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Zone not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a zone by ID
      tags:
      - zones
    put:
      consumes:
      - application/json
      description: Updates an existing zone
      parameters:
      - description: Zone ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated zone
        in: body
        name: zone
        required: true
        schema:
          $ref: '#/definitions/models.Zone'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Zone'
        "400":
          description: Invalid request body or zone ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Zone not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a zone
      tags:
      - zones
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"